+++ tests/next_tests/unit_tests/ffront_tests/test_func_to_foast.py
@@ -203,14 +203,14 @@ def test_binary_pow():
 
 
 def test_binary_mod():
-    def modulo(inp: Field[[TDim], "int64"]):
+    def modulo(inp: Field[[TDim], "int32"]):
         return inp % 3
 
     parsed = FieldOperatorParser.apply_to_function(modulo)
 
     assert parsed.body.stmts[-1].value.type == ts.FieldType(
         dims=[TDim],
-        dtype=ts.ScalarType(kind=ts.ScalarKind.INT64, shape=None),
+        dtype=ts.ScalarType(kind=ts.ScalarKind.int32, shape=None),
     )
